Integrate SteamSpy PySpark job into DAG
